Added ANSI/VT100 support to vt52term and renamed it ansiterm
