+++ apps/build.py
@@ -38,5 +38,5 @@ for prog in [
 
 # Simple C programs.
 
-for prog in ["asm", "copy", "stat", "submit", "objdump", "qe", "life", "vt52term"]:
+for prog in ["asm", "copy", "stat", "submit", "objdump", "qe", "life", "ansiterm"]:
     llvmprogram(name=prog, srcs=["./%s.c" % prog], deps=["lib+cpm65"])
